+++ go.mod
@@ -2,4 +2,7 @@ module Project
 
 go 1.17
 
-require github.com/eclipse/paho.mqtt.golang v1.3.5
+require (
+	github.com/eclipse/paho.mqtt.golang v1.3.5
+	github.com/go-redis/redis/v8 v8.11.4 // indirect
+)
