+++ pkg/mqtt/structs/measures.go
