+++ pkg/mqtt/structs/sensors.go
